Refactors HamCarouselView and enhances name input.
Removes the HamCarouselView's TextField and uses a Text element instead for displaying the name.

Updates HamzzangMainView to incorporate a TextField for name input with focus management. The name is limited to a set number of characters and handles submission and focus events to trigger specific actions upon completion of input. This improves the user experience when setting the name.